test(#615): fold feature96 survivors into downtime-processing.spec.js + delete file
35/50 tests in downtime-processing-feature96.spec.js were retired in #602 as
obsolete (asserted removed ribbon/val-btn subsystems). The 15 survivors — roll
buttons, Confirm Dice Pool states, auto-merit compact panel — are real coverage
but split across a 70%-skipped shell. Fold all 15 into the maintained suite,
add 6 required fixtures (PENDING/CONFIRMED/ROLLED/VALIDATED project states,
FEEDING_PENDING, AUTO_MERIT_PENDING), and delete the feature96 file.

28/28 tests pass; 0 product code changed.
